Switch to file

+++ 01-rocket-equation/rocket_test.go
@@ -4,7 +4,7 @@ import "testing"
 
 func TestFuelForMass(t *testing.T) {
 	cases := []struct {
-		input, expected uint
+		input, expected uint64
 	}{
 		{12, 2},
 		{14, 2},
